Organized vimrc settings
All the settings for the vim and its plugins will be in the vimrc file
and a table of content will be present at the top. Jumps with the help
of '*' (asterick) will be used to jump to the respective section of the
vimrc and use movement keys like 'h j k l <C-u> <C-d>' to move around
and '<C-o> <C-i>' to jump back and forth.

This is experimental and we will see how it goes.