feat(netrw): Trying netrw in place of NERDTree
NERDTree is quite slow and also not going to be my main medium of moving
from file to file. I thought of using nnn[1] as I already use it as my
main terminal file manager but that completely sucked.

Also, I integrated fzf and thus a heavy file explorer does not seem to
be worth it. Trying out the built in netrw[2] with custom mappings
inside the netrw buffer.

In the future, I might try vim-dirvish[3] and fern.vim[4]

[1] https://github.com/jarun/nnn
[2] https://shapeshed.com/vim-netrw/
[3] https://github.com/justinmk/vim-dirvish
[4] https://github.com/lambdalisue/fern.vim