refactor(pyenv): Remove pyenv virtualenv init
`pyenv virtualenv init -` injects a hook in the $PROMPT_COMMAND which
helps in automatically activate the virtual enviroment with the help of
a .python-version file and also enter the respective shell.

The check actually slows down the prompt rendering a bit such that there
are spaces when pressed <Enter> multiple times.