fix(dashboard): correct index for the entries
As we're using utils.append to write lines in a buffer, that will add a
new line after each append, thus subtracting 1 to get the correct index
for the entry line.

- Currently the cursor it at line 1
1. |

- call utils.append with some line {'hello world'}
1. hello world
2. |

Now, the cursor ended up at line two, thus making the total number of
lines in the buffer 2. But our entry line is one above the total.