refactor(icons): get as and when required + cache them
Now, requiring the core icons module will return a metatable which can
be indexed to get the required icon. The process is as follows:

If the icon is present in the cache/store, return it otherwise, get the
icon through its alias or real name and store it for future use in the
icons table.