fix: stop old timer for clearing command messages
The autocmd is triggered on two events: `CmdlineChanged` and `CmdlineLeave`.
So, after typing the command and executing it, thus leaving the cmdline,
if the time is less than `timeout`, the first timer will execute the
callback thus clearing out the messages in short amount of time.

This solves the above inconsistency by resetting the timer.