refactor: updates for Format class:
Write the changes only if they differ from the original input. This is
using `vim.deep_equal` to check whether the original lines are the same
as that of the formatted lines.

Close the handles safely by checking `handle:is_closing()`.
Ref: `plenary.job.close_safely`

Extract out the `stdout` and `stderr` reading steps into a function
which concatenates the chunks and assigns it to the given key.