feat(format): integrate with lsp provided formatting
There will be an option to register for the filetype `use_lsp` which
denotes that we have to use the LSP provided formatter instead. This
will check whether formatting is enabled for the current buffer or not
as usual and then perform the formatting.

LSP has formattingOptions which can be provided with `lsp_opts` key. By
default it will use the ones provided by `vim.lsp.util.make_formatting_params`
but keep the ones provided by the user.