feat(lsp): fallback to treesitter to preview location
Every server can either return a Location or LocationLink where the
former provides only the main location info while the latter provides
the location range for the full body of function, class, etc.

In any case, we will use treesitter to get the location range of the
entire function or class body and even more according to the language.

The variable `node_types` will contain the node types which should be
used from treesitter.

Improvements from the referenced module:
- Use treesitter to highlight the floating window
- Traverse the tree upwards for the same line upto the preferred node type
- Simplify the getting the parser code

Ref: https://github.com/fsouza/dotfiles/blob/main/nvim/lua/fsouza/lsp/locations.lua