style(format): run stylua with new option
> no_call_parenthesis = true

With the new option, we will adopt the lua stylua of having optional
parenthesis for single argument function calls.

Although, looking at the diff of the formatting, some things seems a bit
confusing, like:

-g.startify_session_dir = vim.fn.stdpath("data") .. "/session"
+g.startify_session_dir = vim.fn.stdpath "data" .. "/session"
                                        ^^^^^^^^^^^^^^^^^^^^^

-        venvPath = os.getenv("HOME") .. "/.pyenv",
+        venvPath = os.getenv "HOME" .. "/.pyenv",
                            ^^^^^^^^^^^^^^^^^^^^^^

Concatenate and then call or call and then concatenate? It is understood
that call is taking precedence over concatenating but can someone really
understand that without running the code or looking at the
documentation.

Another is one or two function calls surrounding a binary op:

-      and fn.bufnr("#") ~= fn.bufnr("%")
+      and fn.bufnr "#" ~= fn.bufnr "%"
                   ^^^^^^^^^^^^^^^^^^^^

-  if fn.line2byte("$") ~= -1 then
+  if fn.line2byte "$" ~= -1 then
                  ^^^^^^^^^^^

Are we comparing the string with the output of function call/number?
Again, knowing that function is taking precedence helps but to a
beginner this would be confusing.

Another thing I find interesting is that this change can sometimes be
seen as similar to some scripting languages like ruby. These seems
similar to giving a command an argument instead of calling a function
with a argument. Some examples might help visualize this:

-if vim.fn.executable("rg") then
+if vim.fn.executable "rg" then

-    cmd("packadd packer.nvim")
+    cmd "packadd packer.nvim"