feat(formatter): logic for the daemon server
Update in the formatter spec for the daemon_client:
- cmd: server start command
- args: arguments for the command which should be a function where the
  host and the port number is passed which should be used in the
  arguments
- headers: for options which are passed to the formatter

The process is as follows:

1. The server is registered to be started on a autocmd for the
   respective filetype. A cleanup function is setup which will be called
   before we leave vim to kill all the servers
2. Once the server is started, the state is changed which is available
   as a private field in `formatter._state`
3. On save, the formatting requests is sent to the server using
   `plenary.curl` and the response_handler is used to process the
   response as per the status code

Response handler is different for every formatter as we don't know what
the server will respond. If I find any generic way then I will update
it.

This is still WIP and I just did it for fun. I might not use this at all.