chore: remove unnecessary files
* NOTES.md:

I thought I would write down some notes regarding individual
components in one file, but I think it is better to write them down in
a README file in the respective directory. Eg., config/vimium/README.md,
config/nvim/README.md.

* TODO.md:

I am not using this file that much as I hoped. There's not much TODO in
terms of dotfiles. Yeah, I have some ideas now and then but most of them
are either small that I implement them right away or I write it down on
a paper where my real TODO resides.

* Vagrantfile:

I was using this to experiment with other's dotfiles but now I just look
at the source code. It's also a hassle setting it up although there are
easier ways to automate them but I am not using it that much to
configure it. Maybe I will in the future.