refactor: upgrade neovim structure for lua runtime
As mentioned in the README, the structure have been updated as proposed.

There is a major change in the plugins file where now the configuration
is done manually as described in the packer help.

Also, instead of function, there is now a simple plugins table which is
passed to `packer.use` and is sorted alphabetically for simplicity.

The major benefit of having lua files getting sourced from runtime is
that we can now know the time it took to source a file from
`vim --startuptime` although I use a plugin for better formatting.

The new structure looks something like this:

- init.lua -> sets global variables and sources the globals module
- config/plugin/* -> editor related modules such as autocmds, commands,
  mappings, options, statusline, etc.
- config/after/plugin/* -> plugin configurations
- config/lua/dm/plugin/* -> lazily loaded modules which contains a
  `setup` function
- config/lua/dm/lsp/* -> LSP module

If a plugin uses the `setup` pattern for configuration and that plugin
is lazy loaded then that is put under config/lua/dm/plugin/* as we
cannot put it in runtime path as the plugin is not yet loaded.