feat(packer): custom packer setup for lazy loading
Rationale: There's no point in loading the plugin manager when we don't
actually want to install, update or do anything related to plugins.

Setup:
Commands similar to packer will be defined in a runtime file and other
necessary keybindings and autocmds. When any packer command is invoked,
the plugin and the specs are loaded and then the command is performed
through packer.

Now, the plugin information table will be stored in a file located in
the same directory as that of the compiled file. These are in runtime
paths so will be loaded on startup.

The files will be missing only on a fresh setup and the user will need
to invoke 'PackerInstall' to initiate everything.

Variable rename: _CachedPluginInfo -> _PackerPluginInfo

Ref:
- https://github.com/glepnir/nvim/blob/master/lua/core/pack.lua
- https://github.com/wbthomason/dotfiles/blob/linux/neovim/.config/nvim/lua/plugins.lua