refactor(gs): provide description instead of url
Before:
=======
GitHub Stars > _
<repository name>           <url>
...

After:
======
GitHub Stars > _
<repository name>           <description>

Unlike telescope, fzf cannot act on any data other than what's provided
as the stdin, so I had to provide the url to the second column and use
that with `open`. But, we can use the repository name to construct the
url and thus use the second column for more useful information such as
the description.