fix: reload plugins module before every `Packer*` cmd
When we `require` in lua, it will cache the loaded content and reuse it
for every subsequent call to `require`. Thus, if we were to make any
changes such as adding/removing a plugin, we would have to reload Neovim
to get the necessary updates.

So, we will only reload the module instead. `R` will clear the cache and
then `require` the module.

`R` seems a bit ambiguous, so maybe I will change it something which is
meaningful like `rerequire`?