chore(cheat40): remove unnecessary keymap
As I removed the keymap of 'q:' and 'q/' from telescope, there's no need
to schedule the keybinding and let the plugin define it.

I did this because the plugin would define it without `<nowait>` and so
it would be waiting as there were the above two keymaps. Scheduling it
to a later point solved this issue, but now there's no need.