refactor(lsp): use custom function for lsp messages
Remove the usage of lsp-status plugin for messages as Neovim already
provides the same API internally. The process works as follows:

On event 'User LspProgressUpdate', a function is called which sets the
global variable `lsp_progress_message` to reflect the current message received
from the server. It will also start a timer to trigger a function which will
clear the variable so that the statusline is cleared. This is
debounced to avoid flickering in the statusline.

The statusline will use the above variable to set the message with the
appropriate highlight.