chore: remove homebrew neovim and use local build
As we are on nightly, sometimes Homebrew errors out due to some
dependencies or something else not being updated accordingly. I could
wait a day or two for Homebrew to catch up but meh, like I already have
a local build so why not use that only? Having two nightly builds seems
too much.

In the future, if I want to move to a stable version, I can do that
easily and keep the nightly around as well.