feat(lir): open floating window w.r.t current window
With the new commit (https://github.com/tamago324/lir.nvim/commit/717c379d6339dcb21681e1b46a218df0e184c038), we can now construct
the Lir floating window options dynamically. So, we will construct the
position, width and height of the window such that it is centered only
for the current window and not the editor.

This will be similar to how it would be if I was using plain buffer as
the file explorer.

Thinking more about it, it could be better if I was using the plain
buffer instead of floating window as I could keep the buffer open.
Although I can keep the floating window open as well, so it seems equal.