refactor: back to using brew dump for Brewfile
We will also use the `--describe` flag to add the desription for the
respective package just above the declaration for posterity.

I install and uninstall a lot of packages and I don't want to have a
mental overhead of keeping the file in sync with the current state of my
machine. Let Homebrew do the job.