refactor(mappings): don't keep cursor in the center
I don't remember the reason why I did this, but now I don't like it that
much because as soon as I press any of the keybindings, the entire view
is shifted if the cursor is not already in the center.

I am putting this in a separate commit so that, in case I don't like it,
I can revert it back.