refactor: move border config to global `dm` namespace
Borders are used throughout the config so it makes sense to just include
them in the global namespace as there's only one custom style and the
rest are from Neovim.

Also, I am thinking of hardcoding the icons to the global namespace as
well because there's not many and half are already hardcoded in the
`dm.icons` module. So, move it to `dm.icons` namespace.