feat(telescope): extension to display and copy icon
For now, it only displays icons from the codicons and nonicons set but
may expand to include more. These are around 500 icons and I think it is
enough and there might not be any need to expand this.

Also, this means there's no need for the 'print nonicons icons' code.