refactor(mappings): open URL/directory externally
Use 'gx' to open URL in the browser, similar to netrw which I've
disabled. Similar mapping of 'gx' to open current directory in the
finder from lir buffer.

I'm removing the `external-editor` mapping as I haven't used it ever and
don't think I ever will and so, use that mapping for `external-explorer`
due to similar mnemonic (ee).