refactor: add/update mappings and description
Mappings regarding save and quit will use 'update' which will save the
buffer only if it is changed.

For line movements, '^' and '0' will move on physical/visual lines, not
the actual lines which could be wrapped.

Use 'g_' to jump to end of the line as this will jump to non-blank
character of the line.

On yanking (visual) and pasting (normal + visual), jump to the end of
the yanked or pasted text. It was annoying sometimes when I would select
a few lines and then yank it and the cursor would go right at the
beginning.

Remove split with leader bindings as I directly use the `<C-w>*`
bindings and as such I don't do horizontal splits that much.

Bind 'c' and 'C' similar to 'x' and 'X' which would override what was in
the clipboard.

Change the case for the word under cursor to upper/title case in insert
mode.