feat: update kitty to use FireCode font
Previously, the font was broken on kitty as it would be rendered on the
top position of the row. With the new update comes an option to adjust
the baseline from where the rendering starts.

Other changes reflect the font changes to position it correctly on my
MacBook and monitor.

For some reason, the shell prompt prefix was not being rendered. The fix
was to remove the bold white color which was being used for the icon.

Ref: https://sw.kovidgoyal.net/kitty/conf.html#opt-kitty.adjust_baseline