feat: custom code action lightbulb extmark
Removing another dependency as this is as simple as defining one
function to handle removing and adding extmarks.

The benefit is that I can customize it to however I like and don't have
to depend on the implementation of a plugin.

Also, currently the extmarks are not being highlighted, don't know
what's up with that.