refactor(dashboard): description should be a string
Previously, it was suppose to be a list of strings as functions used to
center the entries and set the lines in the buffer requires a string
table.

But, as description is only of oneline, it makes more sense to keep it
as string and after all the basic operations are performed such as
resolving from a function, adding a key, we can convert the description
to a table and then pass it to center and set it in the buffer.

I have also extracted out some common defaults on the top of the file so
that it is easy to change them when necessary:
- line_length: length of entry line which includes the description and key
- highlights: for header, entry and footer