feat: custom session management module
This is based on vim-startify but a simpler alternative and some
sensible defaults.

All the commands will require one argument, the session name, unlike
startify which prompted the user if no arguments were provided. Also,
command cannot be combined with the bang (!). This simplifies the
functionality and removes lot of lines of code.

There were a few other changes to be made:
- Remove `vim-startify` plugin
- Replace `startify_sessions` with `sessions` for Telescope
- Use `vim.loop.fs_stat` to get the last saved session
- Remove 'options' from sessionoptions