feat: add lua native functions for setting keybindings
API: Each function accepts a table of positional and keyword arguments,
the first and second argument should be the LHS and RHS part of the
mapping and the rest of the arguments should be keyword arguments to
define properties such as <buffer>, <silent>, etc.

This will also accept lua function which will use the `_store` and
`_execute` API functions.

If the user provided a lua function, the conversion will depend on
whether <expr> is set or not. If it is set, then `v:lua` will be use as
vim requires an expression, otherwise the normal way will be used.

Some of the inspirations:
- https://github.com/wincent/wincent
- https://github.com/akinsho/dotfiles