feat: add initial support for snippets
Using LuaSnip.

Also, provide integration with nvim-compe using <tab> and <s-tab> to
either move through the completion menu or move through snippet
placeholders.

The module will be loaded when it is required from compe configuration.
This is something new I learned regarding packer, where the `module` key
can be used to define lazy loading of a plugin based on when the module
is `required`.

Also, started to migrate keybindings to use the native functions.