fix: use index as id for global callbacks
I hesitated to use plain numbers as the id for the global callbacks but
now I am in a situation where my solution breaks :(

While refactoring Telescope mappings to use the new lua native
functions, all the reference pointed to the same location in
`builtin.init.lua` of telescope. This means for every key, it will
override the function even though they are actually different this time.

So, I am going to take the simpler route instead of trying to find some
other way (I tried).