fix: avoid lazy loading treesitter
Ok, this is a long one.

Treesitter module system:
Each module needs to be registered with treesitter using
`treesitter.define_modules()` and most of the modules are defined from
the `plugin/` directory which is loaded on startup. This means if the
treesitter module is lazy loaded, the modules needs to be as well and so
I did. What happened next was completely unexpected. The setup was
something like this:

{ <treesitter>, wants = { <module1>, <module2> }, <lazy loading config> },
{ <module1>, opt = true },
{ <module2>, opt = true },
...

The wants key will load those module before loading the main module
which is what I want. Now, when packer does `packadd <module>`, there is
this error:

> loop or previous error loading module 'nvim-treesitter-textobjects'

which is mostly because the module is loading the module itself, but I
tried to debug and I was not able to find why. So, we are here removing
the lazy loading as I got frustrated by it. I will see what I can do
when I have some time.