fix(lsp): avoid starting language server on preview
When the preview request is made, we are using treesitter to get the
full node at the position. This will result in giving us the entire
node body instead of only the definition.

But, this resulted in the language server being attached because of the
FileType event. We will ignore all events when trying to get the parser
for the buffer.

Also, we will use treesitter to highlight the preview buffer but this
also resulted in a complexity. The mechanism for all LSP floating window
is that pressing one time will open it and another time will put the
cursor in that window. As we are attaching the highlighter to the
preview buffer, trying to focus will try to attach the highlighter again
and result in an error. We will avoid this by setting out own buffer
variable which we will check before making the preview request.

Another thing is that the buffer which we used to get the node using
treesitter is loaded, but unlisted. This means after previewing the
function, if we try to goto the definition of the function, the same
buffer will be opened, as it is already loaded, but the problem is that
there will be no syntax highlighting nor any LSP goodness because we
purposely avoided that. To workaround this, we will delete the buffer
at the end.

That was a lot of workarounds but it seems worth it. Fingers crossed :)