feat: enhance `dm.command` with all builtin attributes
The signature is changed to using arguments instead of table directly.
FROM: `function { "<name>", function/string, <options> }`
TO:   `function("<name>", function/string, { <options> })`

There were a couple of iterations and refactoring which led me to this
simple function design.

First, I updated the function signature as mentioned above. That removed
the code to split the table into positional and keyword arguments.

Then, I had another function `resolve_attr` which would also validate
the type of each attributes, but that seemed unnecessary as vim will
error out if the type or the value is incorrect.

Similarly, there's also no need for setting the default value for
`nargs` to be 0 as vim does it by default.

Refactoring for the existing commands will be done in the next commit.