refactor(commands): use the new interface (4fafd88)
---
Unrelated to this commit but just a thought which poped up while
writing this commit:
What's the difference between an API and an interface?

The things I have written `command`, `autocmd`, etc. are an interface
between lua and vim world, it's a wrapper used to define a vim command
using the lua language. API would be a set of rules defined to talk
with a specific application.

A good read: https://www.ibm.com/cloud/learn/api