REFACTOR: Update map function signature
This change is similar to that for `dm.command` and is done for
similar reason as mentioned in that commit (4fafd88) message.

This will also allow to consolidate options in one variable and use it
to all the similar mappings which will be done in the next commit.

There's also one bug where visual mode mapping is not executed
properly when rhs is a function. This would most probably be because
it is prefixed with a `<Cmd>`. I will investigate soon and send a fix
for it.

The signature is also similar to how Neovim does with the map API
functions where the last argument is for the options.