fix: remove logic for vertical fugitive
The logic was incorrect and then I thought about doing the commit window
in a vertical split for horizontal fugitive but that quickly spiraled
into a lot of possible cases and things got complicated.

So, I just thought best to remove the entire logic as it's not worth
spending time on for the long run.