feat(fugitive): vertical integration
So, I couldn't help it and found a way to integrate fugitive to a
vertical split.

My previous attempts (3f884ab, 01abde2, ec56054, caf6b47, e62e547) had
bad heuristics as I was using number of windows in the current tabpage.

This time I am using the exact layout of the windows in the tabpage and
other heuristics to determine whether the fugitive window needs to be
relocated in a vertical space.