perf(telescope): lazy load github stars extension
I was loading the extension to start the job of collecting the GitHub
stars. This would also load up the required telescope modules which
would take around 20ms to load.

Now, we will start the job when I actually want to search. This hardly
takes 2 seconds, so it's acceptable. I am also thinking about keeping a
state which will make sure to start the job if the information is not
collected yet. This might happen if the internet was not active at the
moment.