perf(telescope): use string keymaps for builtins
Loading `telescope.builtin` takes around 20-30ms to load and so I will
load it only when required. This can either be achieved by creating
thunks (wrapping a call in a function) or just using strings for keymaps.
I am preferring strings for now, but might use thunks if found more
beneficial.

Thunks would've been beneficial if the configuration for the picker
would have to be done at the call site, but now the configuration for
builtin pickers is being done in the setup function.