feat(dashboard): hide the cursor and refactor
The cursor will be hidden in the dashboard buffer as there is no use for
it, everything is done through keymaps. The cursor will be hidden only
in the dashboard buffer and it will be shown when on the commandline or
in any other window/buffer opened through the dashboard entry such as
telescope. The cursor will be resetted after quitting the dashboard.

This means that there is no need to keep the logic for calculating
various cursor position and no autocmds to keep the cursor in the fixed
column within a limit.

This simplifies the code a lot and so I did a bit of refactoring to
group similar statements in one function and made the main function
easier to read (atleast I hope so).