refactor: no bufname for dashboard buffer
I was setting the dashboard buffer name which would make the output for
bufname to be `<directory>/Dashboard` or `~/dotfiles/Dashboard` which is
not correct. This was mainly done to show that we are in a dashboard
buffer in the tabline.

I have now moved that logic in the tabline itself and thus no need to
set the buffer name and also remove the logic to remove the part from
the statusline.