refactor: add commented kitty config file
So, previously I removed all the options which I was not using, but it
feels like going back and forth the documentation a bit inefficient.
Now, all the options will be commented by default and only the overrides
will be uncommented.

Also, the colorscheme config is merged directly in the main config.