refactor: simplify lir buffer keybindings
Create new file/directory with 'f' and 'd' keys instead of 'nf' and 'nd'
as that would create a waiting period if we want to go to the next
instance of the search using 'n'.

'R' to reload.
'cd' to cd (makes sense?)

Rearrange groupwise.

Use metatables for clipboard action which makes it so simpler to use.