refactor: major updates to statusline
I wanted to make it simpler and easy to configure in the future. This
includes not only the components but also the colors. So, the first step
was to simplify the colors and use the bare minimum. Next, using the
User* highlight groups to define those colors and thus making it easier
to incorporate into the statusline.

Now, it was time to simplify the special buffer configuration. I removed
everything and now there exists only a single table which contains the
entire line for the special buffers. One function to be called to get
the line and if it does not exist for the buffer, then return nil.

Remove unnecessary and self explanatory comments.