refactor: merge exports and path into env
After a lot of looking around on how to manage PATH, I've decided to
keep the original implementation and call it a day. I haven't faced any
problem so far, so there's no need to create a problem. Let's see how
far we can go :)

For future me: If this created problems look at:
- https://unix.stackexchange.com/questions/26047/how-to-correctly-add-a-path-to-path/26059#26059
- https://unix.stackexchange.com/questions/88201/whats-the-best-distro-shell-agnostic-way-to-set-environment-variables/88266#88266
- https://superuser.com/questions/187639/zsh-not-hitting-profile/187673#187673
- https://unix.stackexchange.com/questions/609522/should-bashrc-commands-support-shell-reloading