feat(bash): save and reload history
In a typical tmux workflow, there are multiple terminal windows open at
a time and are being used for different purposes. If I have run a
command X in one terminal and I want to run the same command on the
second terminal or maybe something similar, I would have to copy paste
it. This will save the history and reload it everytime the prompt is
drawn.