feat: allow `<Tab>` to be used in Neovim
In the terminal, `<Tab>` and `<C-i>` is seen as the same thing by Vim. To
avoid collision between the two keys, we will map any unused key (`<F6>`)
to `<C-i>` and program the terminal to send `<F6>` on `<C-i>`.

For kitty, it is done using `kitty +kitten show_key` and for other
terminals (Terminal and iTerm2), it is done through Karabiner-elements.

There was a lot of learnings while finding out the key code for <F6>
which I've mentioned thoroughly in `kitty.conf`, so that I don't have to
repeat that again :)

Also, remove spaces between folds in `kitty.conf` and unnecessary file
from Karabiner-elements config folder.