refactor: rearrange tmux config and upgrad copy-mode bindings
Make `v`, `V` and `<C-v>` behave similar to that in Vim. There are some
edge cases but its a good start.

There are other bindings for the prefix table similar to that of Vim
which mostly involes rotating panes, moving panes to far side, etc.

Use folds extensively. I am thinking of adding folds to a lot of places
as now it is easy to toggle them using `<Tab>`.