fix: reset to original position on <esc>
Previously, the behavior was to use the register and restart the search.
This would mean that if after pressing <esc> (leaving search mode), the
cursor would not reset back to the original position from which the
search begin.

Vim already provides a builtin way to move between searches without
leaving incremental search. But, the keys aren't that mnemonic, so let's
remap them appropriately.

`wildcharm` is still needed.