refactor(plugins): load `luasnip` on `InsertEnter`
Luasnip needs to be loaded before nvim-compe as the config for
completion `requires` the module. I could just load it on demand like:

    local luasnip
    ...

    local tab = function()
      luasnip = luasnip or require "luasnip"
      ...
    end

and it works. So, I removed the `wants` keyword and also that is
undocumented as the author wants to unify some keywords.

At the end, our main purpose is to load luasnip before completion plugin
which already happens as the former comes first alphabetically, ignoring
the case.