refactor(telescope): use default mappings as much as possible
`<C-u>` clears the line by default in Vim, so lets just use that instead
of creating a custom action.

This means that the preview scrolling needs to be remapped to something
else: `<C-f>/<C-b>` for forward/backward similar to completion doc.

This will require some change in my muscle memory but that shouldn't be
too hard to adopt to :)