refactor: use pip-tools and separate out packages upgrade
Move from `pipx` to `pip-tools`:

On my way of setting up jupyter, I discovered there could be various
complications with using pipx. There are workarounds by injecting
packages in other packages or installing dependency apps of a package,
but then how would we sync it up with the requirements.txt file? We
could use some complex jq queries or fire up ruby/python to parse and
collect the appropriate metadata but its not worth the effort.

I had an eye on this tool and thought to give it a try as I was moving
to just using global Python version for global packages. This tool not
only keep tracks of dependency but also syncs up the installed packages
with the compiled file. This means that if I remove a package from the
requirements.in then all its transitive packages will be removed as
well.

Another simple refactor is to separate out the packages upgrade process
to python and npm which does what its name suggests (I guess).