feat(bash): function to create ipython kernel in venv
Rationale:

I want jupyter to be installed and available globally but it turns out
that the frontend is different than the kernel. The kernel used in
jupyter will be the one used to install jupyter. This would mean that I
would have to install jupyter in every venv which is not good.

So, I went around the internet and found out that we can create kernel
specs for each venv and tell jupyter to use that kernel from the UI.