refactor(formatter): use new interface and only use stdin
I don't have any formatters which does not take in stdin so there's no
point in keeping that logic around.

If there's a need in the future, it's a simple change of adding two
functions for reading and writing a temp file and passing the filename
along with the formatter args.