fix(diagnostics): do not close on WinScrolled
It seems that the event was closing the preview window prematurely
which was quiet annoying as I had to invoke the function twice to get
the popup window to show. I wondered why was this behavior not
happening in the default implementation and found out the difference.

In addition to that, I have updated the following:
- Resolve diagnostic.code for lsp/linter
- No `showbreak`
- Close window when starting to type
- Use new api to get the diagnostics