feat(diagnostic): separate it out from the lsp module
as now its not related only to the LSP, my linter setup uses the same
api to display the diagnostics.

Also, just use the default implementation as we can format the message
using a function. I still wanted to use icons and proper colors to
separate out the source and code but that meant to maintain another
module.