chore: no rename notification
There are many cases to account for and it's not that worth it. My
main purpose was to know how many files the changes were made so that
I can save them, which can be done easily using `:wall` and to list
the renamed files, just look for "+" in `:ls`.