feat(sh): custom script to manage Python versions
This is similar to `pyenv` except a much smaller version and only
includes a few necessary commands for my need. I only used `pyenv` to
install Python versions and `venv` module to manage virtual environments
so just thought to write this myself.

The structure of the root directory is almost the same except this will
symlink a bin directory to the currently active Python version instead
of going through the shims.