fix(ui): call user function on cancel as well
`lir` started using `vim.ui.input` for the rename action and if I would
cancel the prompt, then the `on_confirm` callback would not be called.
The side effect would be that the current directory would change as the
`rename` function was doing that for completion. So, let the user
provided callback function `on_confirm` handle the cancel part where it
would be called with `nil`.