feat(mac): fn keys for common settings
For termial apps, the touchbar will show function keys and so I need a
way to adjust the common settings like display brightness, volume
control, keyboard backlight. The only issue is I could not figure out
how to have F10-12 work as similar to an external keyboard so I could
use it as volume control and stepping for the debugger.