refactor: enter insert mode only if on the last line
Whenever looking at a long output in the terminal buffer, the automatic
behavior of going into insert mode would be annoying. For example,
looking at the `pytest` output, checking which tests failed and then
going back to the terminal buffer just to scroll down, but no, it would
enter insert mode.

Now, it will enter insert mode only if the cursor is at the last line.