fix(zsh): use absolute path to global python exec
When Python is creating a virtual environment using the builtin venv
module, it creates symlinks to the original executable file. Now, as my
custom `pyenv` script creates a symlink to the bin directory of the
global Python version, the symlink in the venv will also change.

This basically means that when creating a virtual environment from the
global Python version, the venv Python version will always correspond to
the global Python version.