fix: define editorconfig config in plugin spec
The plugin defines a script scoped variable based on the value of the
global variable. Now, that is done in `plugin/editorconfig.vim` file and
I'm defining the variables in `after/` file which means by the time it
picks up my config, the script variables are already set and that will e
used.