dark-mode-notifier: generic mac script
The problem here is that the kitty executable path could be different
depending on whether it's a M-series or intel Mac. I tried adding a
which-like functionality but that didn't work. Then, I realize that
the application path would still be the same and it contains the kitty
executable so why not just use that instead.

Also, let's avoid committing the executable, we'll create it on a new
machine.